+++ packages/vue/src/index.ts
@@ -1,2 +1,4 @@
 export { default as CmButton } from './components/button/CmButton.vue';
 export type { CmButtonProps, CmButtonSize, CmButtonType, CmButtonVariant } from './components/button/button.types';
+export { default as CmCard } from './components/card/CmCard.vue';
+export type { CmCardElement, CmCardProps } from './components/card/card.types';